+++ R/dataParseFunctions.R
@@ -8,23 +8,30 @@ parse_bqr_query <- function(x){
     float = as.double,
     boolean = as.logical,
     string = identity,
-    timestamp = function(x) as.POSIXct(as.integer(x), origin = "1970-01-01", tz = "UTC")
+    timestamp = function(x) as.POSIXct(as.integer(x), origin = "1970-01-01", tz = "UTC"),
+    date = as.Date #fix for #22 if using schema DATE
   )
 
   schema <- x$schema$fields
-  
-  ## when only one row, it makes it 3 obs. of 1 variable instead of 1 obs. of 3
-  data_f <- as.data.frame(Reduce(rbind, lapply(x$rows$f, function(x) x$v)), 
+
+  ## ffs
+  data_f <- as.data.frame(matrix(unlist(unlist(x$rows)), 
+                                 ncol = length(schema$name),
+                                 byrow = TRUE), 
                           stringsAsFactors = FALSE)
   
   types <- tolower(schema$type)
   
-  out <- vector("list", length(types))
-  for(i in seq_along(types)){
-    ## this needs to behave when only length 1
-    out[[i]] <- converter[[types[i]]](data_f[,i])
+  converter_funcs <- converter[types]
+  
+  for(i in seq_along(converter_funcs)){
+    
+    data_f[,i] <- converter_funcs[[i]](data_f[, i])
   }
-  names(out) <- schema$name
+  
+  names(data_f) <- schema$name
+  
+  out <- data_f
   
   out <- as.data.frame(out, stringsAsFactors = FALSE)
   attr(out, "jobReference") <- x$jobReference