+++ tests/testthat.R
@@ -1,0 +1,6 @@
+library(testthat)
+library(googleCloudStorageR)
+library(bigQueryR)
+
+
+test_check("bigQueryR")